% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_line_graphs_workbook.R
\name{create_line_graph}
\alias{create_line_graph}
\alias{function_for_line_graph}
\title{Create line graph}
\usage{
function_for_line_graph(working_directory)
}
\arguments{
\item{working_directory:}{Working directory}
}
\description{
Function that returns a data frame with the total audience by predefined brackets age for a selected interest.
}
\examples{
function_for_line_graph("/Users/user_1/Desktop")
}
\keyword{FB}
\keyword{ads}
\keyword{audience}
