% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dma_ads_breakdown.R
\name{dma_ads_breakdown}
\alias{dma_ads_breakdown}
\title{FB audience media markets breakdown}
\usage{
dma_ads_breakdown(id_vector, name_vector, platform)
}
\arguments{
\item{id_vector:}{Vector of one or more facebook API interests}

\item{name_vector:}{Vector of one or more facebook API names}

\item{platform:}{"instagram", "facebook" or c("facebook", "instagram")}
}
\description{
Function that returns a data frame with the total audience for an interest in each of the top 50 media markets by population
}
\examples{
dma_ads_breakdown("6003123299417", "Computer science", "facebook")
}
\keyword{FB}
\keyword{ads}
\keyword{audience}
