% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dma_comparison.R
\name{dma_comparison}
\alias{dma_comparison}
\title{Creating DMA Comparison}
\usage{
dma_comparison(infile, dma_file)
}
\arguments{
\item{infile:}{.xlsx file containing the audience media market audience}

\item{dma_file:}{.csv file where comparison will be written. Needs to include ".csv" extension}
}
\description{
Function that creates a comparison between audience media market ranking and population media market ranking
}
\examples{
dma_comparison("report_1.xlsx", "dma_comparison.csv")
}
\keyword{Comparison}
\keyword{DMA}
