% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donations.R
\name{donations}
\alias{donations}
\title{FB audience donation breakdown}
\usage{
donations(id_vector, name_vector, platform)
}
\arguments{
\item{id_vector:}{Vector of one or more facebook API interests}

\item{name_vector:}{Vector of one or more facebook API names}

\item{platform:}{"instagram", "facebook" or c("facebook", "instagram")}
}
\description{
Function that returns a data frame with the total audience by donation type for a selected interest.
}
\examples{
donations("6003123299417", "Computer science", "facebook")
}
\keyword{FB}
\keyword{ads}
\keyword{audience}
