% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_preparer.R
\name{excel_preparer}
\alias{excel_preparer}
\title{Create pivot tables for FB audience report}
\usage{
excel_preparer(excel_file)
}
\arguments{
\item{excel_file:}{a user created Facebook audience report}
}
\description{
Function that creates pivot tables and make changes to a user generated report
}
\examples{
excel_preparer("report_1.xlsx")
}
\keyword{audience}
\keyword{excel,}
\keyword{fb}
\keyword{pivot,}
