% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_report.R
\name{retrieve_report}
\alias{retrieve_report}
\title{FB audience generate .xlsx report}
\usage{
retrieve_report(id_vector, name_vector, my_platform, excel_file)
}
\arguments{
\item{id_vector:}{Vector of one or more facebook API interests}

\item{name_vector:}{Vector of one or more facebook API names}

\item{platform:}{"instagram", "facebook" or c("facebook", "instagram")}

\item{excel_file:}{An excel file name, adding ".xlsx" extension is not necessary}
}
\description{
Function that provides a wrapper for most the audience functions and writes an .xlsx report with the data
}
\examples{
retrieve_report("6003123299417", "Computer science", "facebook", "computer_science")
}
\keyword{FB}
\keyword{ads}
\keyword{audience}
