% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_audience.R
\name{us_audience}
\alias{us_audience}
\title{FB audience: Total US Audience}
\usage{
us_audience(id_vector, name_vector, platform)
}
\arguments{
\item{id_vector:}{Vector of one or more facebook API interests}

\item{name_vector:}{Vector of one or more facebook API names}

\item{platform:}{"instagram", "facebook" or c("facebook", "instagram")}
}
\description{
Function that returns a number with the total audience for an interest in the United States
}
\examples{
us_audience("6003123299417", "Computer science", "facebook")
}
\keyword{FB}
\keyword{ads}
\keyword{audience}
